perf(TouchHandler): Improve the performance of touch
Using `IsHitTestVisible`, we can limit the conditions where we need to use `VisualTreeHelper` to only those instances where the first React view in the hit test uses `pointerEvents: box-none`. This should be a reasonably rare case, and can be avoided altogether if you never use a border brush on a view with `pointerEvents: box-none`. Given that any view in a subtree with `pointerEvents: none` at the root will no longer register as the `OriginalSource` for the `PointerRoutedEvent`, we now only need to check whether any parent of the `OriginalSource` applies the `pointerEvents: box-only` to take over as the pointer target.